+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 **master**
 + Properly ignores jobs that do not have `Sidekiq::Status::Worker` included
 + Honors custom job expirations for ActiveJob jobs
++ Adds a `:retrying` status
++ Server middleware will now catches all exception types
++ Changes where server middleware is inserted in the chain
 
 **Version 0.7.0**
 + Sidekiq 4.2 and 5 now supported